% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLabDefinitions.R
\name{getLabDefinitions}
\alias{getLabDefinitions}
\title{Returns a dataframe with lab definitions}
\usage{
getLabDefinitions()
}
\arguments{
}
\value{
A dataframe
}
\description{
Returns a dataframe with lab definitions
}
\details{
Returns a dataframe with lab definitions
}
