% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCharacterizationExport.R
\name{runCharacterizationExport}
\alias{runCharacterizationExport}
\title{runCharacterizationExport}
\usage{
runCharacterizationExport(connectionDetails,
                          cdmDatabaseSchema,
                          targetDatabaseSchema,
                          targetTable,
                          outcomeDatabaseSchema,
                          outcomeTable,
                          outputDirectory,
                          executionPath,
                          csvFilePrefix,
                          databaseId)
}
\arguments{
\item{connectionDetails}{}

\item{targetDatabaseSchema}{}

\item{targetTable}{}

\item{outcomeDatabaseSchema}{}

\item{outcomeTable}{}

\item{outputDirectory}{}

\item{executionPath}{}

\item{csvFilePrefix}{}

\item{databaseId}{}
}
\value{
A dataframe
}
\description{
Run Characterization
}
\details{
Run Characterization
}
