% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runIncidenceAnalysis.R
\name{runIncidenceAnalysis}
\alias{runIncidenceAnalysis}
\title{Run MTXlab CohortIncidence Analysis}
\usage{
runIncidenceAnalysis <- function(
    connectionDetails,
    cdmDatabaseSchema,
    cohortDatabaseSchema,
    cohortTable,
    workDatabaseSchema = cohortDatabaseSchema,
    sourceName,
    refId = 1,
    outputFolder = NULL
}
\arguments{
\item{connectionDetails}{}

\item{cdmDatabaseSchema}{}

\item{cohortDatabaseSchema}{}

\item{cohortTable}{}

\item{workDatabaseSchema}{}

\item{sourceName}{}

\item{refId}{}

\item{outputFolder}{}
}
\value{
A dataframe
}
\description{
Run MTXlab CohortIncidence Analysis
}
\details{
Run MTXlab CohortIncidence Analysis
}
