% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runKMAnalysis.R
\name{runKMAnalysis}
\alias{runKMAnalysis}
\title{runKMAnalysis}
\usage{
runKMAnalysis(connectionDetails,
              cdmDatabaseSchema,
              cdmDatabaseName,
              cohortDatabaseSchema,
              cohortTable,
              outcomeDatabaseSchema,
              outcomeTable,
              outputFolder)
}
\arguments{
\item{connectionDetails}{}

\item{cdmDatabaseSchema}{}

\item{cdmDatabaseName}{}

\item{cohortDatabaseSchema}{}

\item{cohortTable}{}

\item{outcomeDatabaseSchema}{}

\item{outcomeTable}{}

\item{outputFolder}{}

}
\value{
A dataframe
}
\description{
Run KM Analysis
}
\details{
Run KM Analysis
}
