% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runLabFollowUp.R
\name{runLabFollowUp}
\alias{runLabFollowUp}
\title{Calculate Lab Follow Up for all the iterations}
\usage{
runLabFollowUp(con,
               workDatabaseSchema,
               cohortTable,
               cdmDatabaseSchema)
}
\arguments{
\item{con}{}
  
\item{workDatabaseSchema}{}

\item{cohortTable}{}

\item{cdmDatabaseSchema}{}
}
\value{
A dataframe
}
\description{
Calculate Lab Follow Up
}
\details{
Calculate Lab Follow Up
}
