% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateLabFollowUp.R
\name{calculateLabFollowUp}
\alias{calculateLabFollowUp}
\title{Calculate Lab Follow Up}
\usage{
calculateLabFollowUp(con,
                     workDatabaseSchema,
                     cohortTable,
                     cdmDatabaseSchema,
                     labName,
                     cohortDefinitionId,
                     ancestorConceptId,
                     abnormalLowerLimit,
                     abnormalUpperLimit)
}
\arguments{
\item{workDatabaseSchema}{}

\item{cohortTable}{}

\item{cdmDatabaseSchema}{}
}
\value{
A dataframe
}
\description{
Create qrisk 1 model for females using Qrisk cohorts
}
\details{
Calculate Lab Follow Up
}
